Remove old file protocol
