+++ protocols/file/handlers.go
@@ -1,235 +1,259 @@
 package file
 
 import (
-	"bytes"
 	"context"
 	"encoding/binary"
+	"fmt"
 	"io"
-	"io/ioutil"
 
+	"github.com/functionland/go-fula/drive"
+	fxiface "github.com/functionland/go-fula/fxfs/core/iface"
 	files "github.com/ipfs/go-ipfs-files"
-	coreiface "github.com/ipfs/interface-go-ipfs-core"
-	"github.com/ipfs/interface-go-ipfs-core/path"
+
+	"github.com/golang/protobuf/proto"
 	"github.com/libp2p/go-libp2p-core/network"
-	"google.golang.org/protobuf/proto"
 )
 
-func RequestHandler(ctx context.Context, api coreiface.CoreAPI, stream network.Stream) {
-	pref_buf := make([]byte, binary.MaxVarintLen64)
-	pref_reader := bytes.NewReader(pref_buf)
-	n, err := io.ReadFull(stream, pref_buf)
+// Root handler for newFile protocol
+// It reads 2 bytes from the stream which determines the size of the request messeage
+// It then reads the request message and passes the stream to the appropriate handler
+func Handle(ctx context.Context, api fxiface.CoreAPI, ds drive.DriveStore, s network.Stream) error {
+	log.Debug("handling")
+
+	sizebuf := make([]byte, 2)
+	_, err := io.ReadFull(s, sizebuf)
 	if err != nil {
-		log.Error("fail to read perfix", err)
-		stream.Reset()
-		return
+		return err
 	}
-	log.Debug("read %d byte from stream", n)
-	msg_size, err := binary.ReadUvarint(pref_reader)
+
+	reqbuf := make([]byte, binary.LittleEndian.Uint16(sizebuf))
+	_, err = io.ReadAtLeast(s, reqbuf, int(binary.LittleEndian.Uint16(sizebuf)))
 	if err != nil {
-		log.Error("fail to pars perfix", err)
-		stream.Reset()
-		return
+		return err
 	}
-	log.Debug("readed message size from perfix ", msg_size)
-	header, err := io.ReadAll(pref_reader)
-	msg_buf := make([]byte, int(msg_size)-len(header))
-	if err != nil {
-		log.Error("fail to read header", err)
-		stream.Reset()
-		return
-	}
-	req := &Request{}
-	n, err = io.ReadFull(stream, msg_buf)
-	if err != nil {
-		log.Error("fail to read header", err)
-		stream.Reset()
-		return
-	}
-	log.Debug("readed message with size ", n)
-	header = append(header, msg_buf...)
-	err = proto.Unmarshal(header, req)
-	if err != nil {
-		log.Error("can't unmarshal header bytes ", err)
-		stream.Reset()
-		return
-	}
-	switch v := req.Type.(type) {
-	case *Request_Receive:
-		ReceiveFileHandler(v, api, stream, ctx)
-	case *Request_Meta:
-		ReceiveMetaHandler(v, api, stream, ctx)
-	case *Request_Send:
-		SendFileHandler(v, api, stream, ctx)
+
+	req := &FSRequest{}
+	proto.Unmarshal(reqbuf, req)
+
+	switch req.Action {
+	case ActionType_READ:
+		return HandleRead(ctx, api, ds, req.Path, req.DID, s)
+	case ActionType_MKDIR:
+		return HandleMkDir(ctx, api, ds, req.Path, req.DID, s)
+	case ActionType_WRITE:
+		return HandleWrite(ctx, api, ds, req.Path, req.DID, s)
+	case ActionType_DELETE:
+		return HandleDelete(ctx, api, ds, req.Path, req.DID, s)
+	case ActionType_LS:
+		return HandleLs(ctx, api, ds, req.Path, req.DID, s)
 	default:
-		log.Error("message not supported")
+		s.Reset()
+		return fmt.Errorf("Action Type not supported: %s", req.Action)
 	}
 }
 
-func ReceiveFileHandler(req *Request_Receive, api coreiface.CoreAPI, stream network.Stream, ctx context.Context) {
-	file_path := path.New("/ipfs/" + req.Receive.Id)
-	f, err := api.Unixfs().Get(ctx, file_path)
+// Read a file at a given path from the user's drive
+func HandleRead(ctx context.Context, api fxiface.CoreAPI, ds drive.DriveStore, path string, userDID string, s network.Stream) error {
+
+	ud, err := ds.ResolveCreate(userDID)
 	if err != nil {
-		log.Error("cant not resolve file ", err)
-		stream.Reset()
-		return
+		log.Error(err)
+		return err
 	}
-	var fileNode files.File
-	switch f := f.(type) {
-	case files.File:
-		fileNode = f
-	case files.Directory:
-		log.Error("not file")
-		return
-	default:
-		log.Error("not file")
-		return
-	}
-	fBytes, err := ioutil.ReadAll(fileNode)
+	ud.Publish(ctx, api)
+
+	ps, err := ud.PublicSpace(ctx, api)
 	if err != nil {
-		log.Error("error reading file ", err)
-		stream.Reset()
-		return
+		log.Error(err)
+		return err
 	}
-	fileProto := &File{}
-	err = proto.Unmarshal(fBytes, fileProto)
+
+	file, err := ps.ReadFile(path, drive.ReadFileOpts{})
 	if err != nil {
-		log.Error("connot unmarshal file", err)
-		stream.Reset()
-		return
+		log.Error(err)
+		return err
 	}
-	content_path := path.New(fileProto.GetContentPath())
-	c, err := api.Unixfs().Get(ctx, content_path)
-	if err != nil {
-		log.Error("cant not resolve file ", err)
-		stream.Reset()
-		return
-	}
-	var fileNode1 files.File
-	switch c := c.(type) {
-	case files.File:
-		fileNode1 = c
-	case files.Directory:
-		log.Error("not file")
-		return
-	default:
-		log.Error("not file")
-		return
-	}
-	fBytes1, err := ioutil.ReadAll(fileNode1)
-	log.Info("size of out file:  ", len(fBytes1))
-	if err != nil {
-		log.Error("error reading file ", err)
-		stream.Reset()
-		return
-	}
-	n, err := stream.Write(fBytes1)
-	if err != nil {
-		log.Error("can not write to stream ", err)
-		stream.Reset()
-		return
-	}
-	log.Info("write size to stream :  ", n)
-	err = stream.CloseWrite()
-	if err != nil {
-		log.Error("can not close stream ", err)
-		stream.Reset()
-		return
-	}
+
+	go func() {
+		defer s.CloseWrite()
+
+		n, err := io.Copy(s, file)
+		if err != nil {
+			log.Error("Error in writing file on stream", err)
+		}
+
+		log.Infof("wrote %d bytes to the stream", n)
+	}()
+
+	return nil
 }
 
-func ReceiveMetaHandler(req *Request_Meta, api coreiface.CoreAPI, stream network.Stream, ctx context.Context) {
-	path := path.New("/ipfs/" + req.Meta)
-	f, err := api.Unixfs().Get(ctx, path)
+// Create a new directory at a given location in the user's drive
+func HandleMkDir(ctx context.Context, api fxiface.CoreAPI, ds drive.DriveStore, path string, userDID string, s network.Stream) error {
+	ud, err := ds.ResolveCreate(userDID)
 	if err != nil {
-		log.Error("cant not resolve file ", err)
-		stream.Reset()
-		return
+		log.Error(err)
+		return err
 	}
-	var fileNode files.File
-	switch f := f.(type) {
-	case files.File:
-		fileNode = f
-	case files.Directory:
-		log.Error("not file")
-		stream.Reset()
-		return
-	default:
-		log.Error("not file")
-		stream.Reset()
-		return
-	}
-	fBytes, err := ioutil.ReadAll(fileNode)
+	ud.Publish(ctx, api)
+
+	ps, err := ud.PublicSpace(ctx, api)
 	if err != nil {
-		log.Error("error reading file ", err)
-		stream.Reset()
-		return
+		log.Error(err)
+		return err
 	}
-	fileProto := &File{}
-	err = proto.Unmarshal(fBytes, fileProto)
+
+	dcid, err := ps.MkDir(path, drive.MkDirOpts{})
 	if err != nil {
-		log.Error("connot unmarshal file", err)
-		stream.Reset()
-		return
+		log.Error(err)
+		return err
 	}
-	meta, err := proto.Marshal(fileProto.Meta)
+
+	ps.Save()
+	err = ud.Publish(ctx, api)
 	if err != nil {
-		log.Error("connot marshal meta", err)
-		stream.Reset()
-		return
+		return err
 	}
-	n, err := stream.Write(meta)
-	if err != nil {
-		log.Error("can not write to stream", err)
-	}
-	log.Debugf("wrote meta with size %d on stream", n)
-	err = stream.CloseWrite()
-	if err != nil {
-		log.Error("can not close stream ", err)
-		stream.Reset()
-		return
-	}
+
+	go func() {
+		defer s.CloseWrite()
+
+		if _, err := s.Write([]byte(dcid)); err != nil {
+			log.Error("error in writing directory cid on the stream", err)
+		}
+	}()
+
+	return nil
 }
 
-func SendFileHandler(req *Request_Send, api coreiface.CoreAPI, stream network.Stream, ctx context.Context) {
-	fBytes, err := ioutil.ReadAll(stream)
-	log.Info("size of input file:  ", len(fBytes))
+// Handle FSRequest with Delete action, delete a Node at a given location in the user's drive
+func HandleDelete(ctx context.Context, api fxiface.CoreAPI, ds drive.DriveStore, path string, userDID string, s network.Stream) error {
+	ud, err := ds.ResolveCreate(userDID)
 	if err != nil {
-		log.Error("stream error ", err)
-		stream.Reset()
-		return
+		log.Error(err)
+		return err
 	}
-	file_cid, err := api.Unixfs().Add(ctx, files.NewBytesFile(fBytes))
-	log.Debug("file added with cid ", file_cid)
-	if err != nil {
-		log.Error("write file to ipfs failed ", err)
-		stream.Reset()
-		return
-	}
-	fileProto := &File{Meta: req.Send, ContentPath: file_cid.String()}
-	log.Debug("FileLink added", fileProto)
-	fileLinkByte, err := proto.Marshal(fileProto)
-	if err != nil {
-		log.Error("creating proto of file link failed ", err)
-		stream.Reset()
-		return
-	}
-	filelink_cid, err := api.Unixfs().Add(ctx, files.NewBytesFile(fileLinkByte))
-	if err != nil {
-		log.Error("can not write the filelink to ipfs ", err)
-		stream.Reset()
-		return
-	}
-	log.Debug("FileLink added and with cid ", filelink_cid)
+	ud.Publish(ctx, api)
 
-	n, err := stream.Write([]byte(filelink_cid.Cid().Hash().B58String()))
+	ps, err := ud.PublicSpace(ctx, api)
 	if err != nil {
-		log.Error("can not write cid to stream", err)
-		stream.Reset()
+		log.Error(err)
+		return err
 	}
-	log.Debugf("wrote cid with size %d on stream", n)
-	err = stream.CloseWrite()
+
+	_, err = ps.DeleteFile(path, drive.DeleteFileOpts{})
 	if err != nil {
-		log.Error("failed to write cid to stream", err)
-		stream.Reset()
+		log.Error(err)
+		return err
 	}
+
+	ps.Save()
+	err = ud.Publish(ctx, api)
+	if err != nil {
+		return err
+	}
+
+	go func() {
+		defer s.CloseWrite()
+
+		if _, err := s.Write([]byte{1}); err != nil {
+			log.Error("error in writing delete ack on the stream", err)
+		}
+	}()
+
+	return nil
+}
+
+// Handle FSRequest with Ls action, list entries existing in a given path in a user's drive
+func HandleLs(ctx context.Context, api fxiface.CoreAPI, ds drive.DriveStore, path string, userDID string, s network.Stream) error {
+	ud, err := ds.ResolveCreate(userDID)
+	if err != nil {
+		log.Error(err)
+		return err
+	}
+	ud.Publish(ctx, api)
+
+	ps, err := ud.PublicSpace(ctx, api)
+	if err != nil {
+		log.Error(err)
+		return err
+	}
+
+	de, err := ps.ListEntries(path, drive.ListEntriesOpts{})
+	if err != nil {
+		log.Error(err)
+		return err
+	}
+
+	items := make([]*DirEntry, 0)
+	for d := range de {
+		items = append(items, &DirEntry{Name: d.Name, Type: EntryType(d.Type), Cid: d.Cid.String(), Size: int32(d.Size)})
+	}
+	dents := &DirEntries{Items: items}
+
+	go func() {
+		defer s.CloseWrite()
+
+		dbuf, err := proto.Marshal(dents)
+		if err != nil {
+			log.Error("error in encoding the DirEntries")
+		}
+		if _, err := s.Write(dbuf); err != nil {
+			log.Error("error in writing directory entries on the stream", err)
+		}
+	}()
+
+	return nil
+}
+
+// Handle FSRequest with Write action, write a file at a given location in the user's drive
+func HandleWrite(ctx context.Context, api fxiface.CoreAPI, ds drive.DriveStore, path string, userDID string, s network.Stream) error {
+	ud, err := ds.ResolveCreate(userDID)
+	if err != nil {
+		log.Error(err)
+		return err
+	}
+	err = ud.Publish(context.Background(), api)
+	if err != nil {
+		log.Error(err)
+		return err
+	}
+
+	ps, err := ud.PublicSpace(ctx, api)
+	if err != nil {
+		log.Error(err)
+		return err
+	}
+
+	pr, pw := io.Pipe()
+
+	f := files.NewReaderFile(pr)
+
+	go func() {
+		defer s.CloseRead()
+		defer pw.Close()
+
+		io.Copy(pw, s)
+	}()
+
+	fcid, err := ps.WriteFile(path, f, drive.WriteFileOpts{})
+	if err != nil {
+		log.Error("error in receiving file", err)
+	}
+
+	ps.Save()
+	err = ud.Publish(ctx, api)
+	if err != nil {
+		return err
+	}
+
+	go func() {
+		defer s.Close()
+
+		if _, err := s.Write([]byte(fcid)); err != nil {
+			log.Error("error in writing the file cid on the stream", err)
+		}
+	}()
+
+	return nil
 }
