+++ protocols/file/protos/file.proto
@@ -2,26 +2,51 @@ syntax = "proto3";
 option go_package = "github.com/functionland/go-fula/protocols/file";
 package file;
 
-message Meta {
-  string name = 1;
-  string type = 2;
-  uint64 size = 3 [jstype = JS_NUMBER];
-  int64 lastModified = 4 [jstype = JS_NUMBER];
+message FSRequest {
+    string DID = 1;
+    string path = 2;
+    bytes signature = 3;
+    ActionType action = 4;
 }
 
-message Chunk {
-  string id = 1;
+enum ActionType {
+    READ = 0;
+    WRITE = 1;
+    MKDIR = 2;
+    LS = 3;
+    DELETE = 4;
 }
 
-message Request {
-  oneof type {
-    string meta = 1;
-    Chunk receive = 2;
-    Meta send = 3;
-  }
+message Payload {
+    repeated int32 d = 1;
+}
+
+message Response {
+    ActionType action = 1;
+    oneof body {
+        File file = 2;
+        string cid = 3;
+        DirEntries entries = 4;
+    }
+}
+
+enum EntryType {
+    FILE = 0;
+    DIRECTORY = 1;
 }
 
 message File {
-  Meta meta = 1;
-  string contentPath = 2;
+    string name = 1;
+    repeated int32 d = 2;
+}
+
+message DirEntries {
+    repeated DirEntry items = 1;
+}
+
+message DirEntry {
+    string name = 1;
+    EntryType type = 2;
+    int32 size = 3;
+    string cid = 4;
 }
