+++ protocols/file/requests.go
@@ -1,136 +1,147 @@
 package file
 
 import (
+	"context"
 	"encoding/binary"
 	"io"
-	"io/ioutil"
-	"sync"
 
+	"github.com/golang/protobuf/proto"
 	logging "github.com/ipfs/go-log"
 	"github.com/libp2p/go-libp2p-core/network"
-	"google.golang.org/protobuf/proto"
 )
 
-var log = logging.Logger("fula:filePL")
+var log = logging.Logger("newFilePL")
 
-func encodPerfixLenth(b []byte) []byte {
-	perfix := make([]byte, binary.MaxVarintLen64)
-	perfixsize := binary.PutUvarint(perfix, uint64(len(b)))
-	header := append(perfix[:perfixsize], b...)
-	return header
+func sendRequest(req *FSRequest, s network.Stream) error {
+	reqMsg, err := proto.Marshal(req)
+	if err != nil {
+		return err
+	}
+
+	sbuf := make([]byte, 2)
+	binary.LittleEndian.PutUint16(sbuf, uint16(len(reqMsg)))
+	n, err := s.Write(sbuf)
+	if err != nil {
+		return err
+	}
+	log.Debugf("req size sent on the stream, wrote %d bytes on stream", n)
+
+	n, err = s.Write(reqMsg)
+	if err != nil {
+		return err
+	}
+	log.Debugf("req sent, wrote %d bytes on stream", n)
+
+	return nil
 }
 
-func ReceiveFile(stream network.Stream, cid string) (io.Reader, error) {
-	reqMsg := &Request{Type: &Request_Receive{Receive: &Chunk{Id: cid}}}
-	header, err := proto.Marshal(reqMsg)
-	if err != nil {
-		log.Error("can not create request message")
-		return nil, err
-	}
-	log.Debug("request message created")
-	header = encodPerfixLenth(header)
-	_, err = stream.Write(header)
-	if err != nil {
-		log.Error("sending Request Message failed", err)
-		return nil, err
-	}
-	log.Debug("request message sent")
-	stream.CloseWrite()
-	return stream, nil
-}
+// Send an FSRequest for Read action, return a io.Reader to read file from the stream
+func RequestRead(ctx context.Context, s network.Stream, path string, userDID string) (io.Reader, error) {
 
-func ReceiveMeta(stream network.Stream, cid string) ([]byte, error) {
-	reqMsg := &Request{
-		Type: &Request_Meta{Meta: cid}}
+	fsReq := &FSRequest{DID: userDID, Path: path, Action: ActionType_READ}
 
-	header, err := proto.Marshal(reqMsg)
+	err := sendRequest(fsReq, s)
 	if err != nil {
-		log.Error("marshaling Meta Message failed", err)
-		return nil, err
-	}
-	header = encodPerfixLenth(header)
-	n, err := stream.Write(header)
-	if err != nil {
-		log.Error("write header faild", err)
 		return nil, err
 	}
 
-	log.Debugf("wrote meta header with size of %d on stream", n)
-	err = stream.CloseWrite()
-	if err != nil {
-		log.Error("failed to close write sream with error: ", err)
-		return nil, err
-	}
-	buf, err := ioutil.ReadAll(stream)
-	if err != nil {
-		log.Error("failed to read Meta buffer. ", err)
-		return nil, err
-	}
-	err = stream.CloseRead()
-	if err != nil {
-		log.Error("failed to close read stream. ", err)
-		stream.Reset()
-		return nil, err
-	}
-	stream.Close()
-	return buf, nil
-}
+	pr, pw := io.Pipe()
 
-func SendFile(fileCh <-chan []byte, filemeta *Meta, stream network.Stream, wg *sync.WaitGroup) (*string, error) {
-	//create header message
-	reqMsg := &Request{Type: &Request_Send{Send: filemeta}}
-	msg, err := proto.Marshal(reqMsg)
-	if err != nil {
-		return nil, err
-	}
-	header := encodPerfixLenth(msg)
-	n, err := stream.Write(header)
-	log.Debugf("header of size %d wrote on stream", n)
-	if err != nil {
-		log.Debugf("failed to wrote header on stream ", err)
-		stream.Reset()
-		return nil, err
-	}
-	log.Debug("request message send")
+	go func() {
+		defer pw.Close()
+		defer s.Close()
 
-	//write file channel to stream
-	for res := range fileCh {
-		log.Debug("file protocol channel data: ", res[:10])
-		n, err := stream.Write(res)
-		if err != nil {
-			log.Error("can not write on stream ", err)
-			stream.Reset()
-			return nil, err
+		if _, err := io.Copy(pw, s); err != nil {
+			log.Error("error in reading file from stream", err)
 		}
-		log.Debugf("write %d on stream", n)
-		wg.Done()
-	}
-	log.Debug("file sent")
+	}()
 
-	//closing wtite stream
-	err = stream.CloseWrite()
-	if err != nil {
-		log.Error("can't close write stream. ", err)
-		stream.Reset()
-		return nil, err
-	}
-	log.Debug("stream write closed succsesfuly")
+	return pr, nil
+}
 
-	//reading stream for cid
-	buf2, err := ioutil.ReadAll(stream)
+// Send an FSRequest for MkDir action, wait until receiving a cid for the directory
+func RequestMkDir(ctx context.Context, s network.Stream, path string, userDID string) (string, error) {
+	fsReq := &FSRequest{DID: userDID, Path: path, Action: ActionType_MKDIR}
+
+	err := sendRequest(fsReq, s)
 	if err != nil {
-		log.Error("cant read cid from stream.")
-		stream.Reset()
+		return "", err
+	}
+
+	bcid, err := io.ReadAll(s)
+	if err != nil {
+		log.Error("coudn't receive the cid ", err)
+		s.Reset()
+		return "", err
+	}
+
+	return string(bcid), nil
+}
+
+// Send an FSRequest for Write action, wait until receiving a cid for the file
+func RequestWrite(ctx context.Context, s network.Stream, path string, userDID string, f io.Reader) (string, error) {
+	fsReq := &FSRequest{DID: userDID, Path: path, Action: ActionType_WRITE}
+
+	err := sendRequest(fsReq, s)
+	if err != nil {
+		return "", err
+	}
+
+	if _, err := io.Copy(s, f); err != nil {
+		log.Error("error writing the file on the stream: ", err)
+		return "", err
+	}
+	log.Debug("wrote file on the stream successfuly")
+	s.CloseWrite()
+
+	bcid, err := io.ReadAll(s)
+	if err != nil {
+		log.Error("coudn't receive the cid ", err)
+		s.Reset()
+		return "", err
+	}
+	s.CloseRead()
+
+	return string(bcid), nil
+}
+
+// Send and FSRequest for Delete action, wait until receiving an ack (a one byte)
+func RequestDelete(ctx context.Context, s network.Stream, path string, userDID string) error {
+	fsReq := &FSRequest{DID: userDID, Path: path, Action: ActionType_DELETE}
+
+	err := sendRequest(fsReq, s)
+	if err != nil {
+		return err
+	}
+
+	_, err = io.ReadAll(s)
+	if err != nil {
+		log.Error("couldn't receive the cid ", err)
+		s.Reset()
+		return err
+	}
+
+	return nil
+}
+
+// Send and FSRequest for Ls action, wait until receiving a list of DirEntries
+func RequestLs(ctx context.Context, s network.Stream, path string, userDID string) (*DirEntries, error) {
+	fsReq := &FSRequest{DID: userDID, Path: path, Action: ActionType_LS}
+
+	err := sendRequest(fsReq, s)
+	if err != nil {
 		return nil, err
 	}
-	id := string(buf2)
-	log.Debugf("received cid: %s", id)
-	err = stream.CloseRead()
+
+	dbuf, err := io.ReadAll(s)
 	if err != nil {
-		log.Error("failed to close read stream. ", err)
-		stream.Reset()
+		log.Error("couldn't receive the cid ", err)
+		s.Reset()
 		return nil, err
 	}
-	stream.Close()
-	return &id, nil
+
+	dentries := &DirEntries{}
+	proto.Unmarshal(dbuf, dentries)
+
+	return dentries, nil
 }
