--- protocols/newFile/constant.go
@@ -1,3 +1,0 @@
-package newFile
-
-const ProtocolId = "/fx/new-file/0.1.0"
